Refactoring explorer model code
