+++ app/adapters/acs-row.js
@@ -7,8 +7,8 @@ const { SupportServiceHost } = Environment;
 
 export default DS.JSONAPIAdapter.extend({
   query(store, modelType, query) {
-    const { selectionId, comparator = 0 } = query;
-    const URL = `${SupportServiceHost}/survey/acs/${selectionId}?compare=${comparator}`;
+    const { selectionId, compareTo = 0 } = query;
+    const URL = `${SupportServiceHost}/survey/acs/${selectionId}?compareTo=${compareTo}`;
 
     return fetch(URL)
       .then(d => d.json());
